+++ packages/hoppity/src/index.ts
@@ -2,6 +2,7 @@ import hoppity from "./hoppity";
 import {
     BrokerCreatedCallback,
     BuilderInterface,
+    Hoppity,
     MiddlewareFunction,
     MiddlewareResult,
     MiddlewareContext,
@@ -15,6 +16,7 @@ export default hoppity;
 export type {
     BrokerCreatedCallback,
     BuilderInterface,
+    Hoppity,
     MiddlewareFunction,
     MiddlewareResult,
     MiddlewareContext,