+++ packages/hoppity-logger/CHANGELOG.md
@@ -1,5 +1,16 @@
 # @apogeelabs/hoppity-logger
 
+## 0.2.0
+
+### Minor Changes
+
+- Ready for beta testing
+
+### Patch Changes
+
+- Updated dependencies
+    - @apogeelabs/hoppity@0.2.0
+
 ## 0.1.0
 
 ### Minor Changes